refactor: ExampleDialog extracted to own file
